Fixed rating and added admin role + view

+++ UnblockMe/UnblockMe/Models/Ratings.cs
@@ -10,7 +10,7 @@ namespace UnblockMe.Models
         public string rater_id { get; set; }
         public string rated_id { get; set; }
         public string rating_message { get; set; }
-        public int rating_value { get; set; }
+        public double rating_value { get; set; }
         public virtual Users Rater { get; set; }
         public virtual Users Rated { get; set; }
     }
